Implement speed limits for second-order system
We don't require the second-order system to use the maximum acceleration
at every timestep. (Doing so actually reduces maneuverability.)

Due to time limitations of the MPPI optimization, we also introduce a
heuristic "stop" behavior that uses the robot's current velocity to
decide which controls will bring the robot to a stop.